Functionality.
Clicking actually implements things. Try it out on the control center
page. The personal money is also increased. Five of your clicks makes a
dollar. Total group money is unchanged as I figured we could do that
live from the backend. On the upgrade screen, upgrades are only shown as
available if you can purchase them. This is dynamic, so if you have
enough clicks for an upgrade, it'll tell you. Fun fun.